02 save msg

+++ day05/ex02/Chat/src/main/resources/schema.sql
@@ -3,21 +3,21 @@ drop schema if exists chat cascade;
 create schema if not exists chat;
 
 create table if not exists chat.user (
-    id          serial primary key,
+    id          bigserial primary key,
     login       varchar(50) unique not null,
     password    varchar(50) not null
 );
 
 create table if not exists chat.chat_room (
-    id          serial primary key,
+    id          bigserial primary key,
     name        varchar(50) unique not null,
-    owner       int not null references chat.user (id)
+    owner       bigint not null references chat.user (id)
 );
 
 create table if not exists chat.message (
-    id          serial primary key,
-    author      integer not null references chat.user (id),
-    room        integer not null references chat.chat_room (id),
+    id          bigserial primary key,
+    author      bigint not null references chat.user (id),
+    room        bigint not null references chat.chat_room (id),
     text        text not null,
     date_time   timestamp default CURRENT_TIMESTAMP
 );
